Delete FAQ.ca.md
This is a duplicate of the file faq.ca.md.  When visitors go to the FAQ page ( http://votersedge.org/ca/page/faq ) they get the other file.  The data in this file is out of date, and the name collision with the lowercase faq.ca.md on Windows systems makes it impossible for Allie and me to work in this repo locally.